Make sure that "release" is always quoted as mysql believes it to be a reserved word.  Store the floats that time.time() gives us as doubles in mysql otherwise sessions don't work very well.
git-svn-id: http://current.tigris.org/svn/current/trunk@279 7112b371-5efc-0310-9db1-efd4ebc68ee9

+++ current/db/mysql/schema.py
@@ -92,8 +92,8 @@ drop table if exists SESSIONS;
 create table SESSIONS (
    session_id     INTEGER PRIMARY KEY auto_increment,
    sid            varchar(40) unique not null,
-   createtime     float not null,
-   timeout        float not null,
+   createtime     double not null,
+   timeout        double not null,
    data           text,
 
    index(sid)
